android sample 0.1
