+++ Android/.idea/.name
@@ -1,1 +1,1 @@
-ImitationDYShortVideo
+AgoraMultiChannelSwitch
