+++ Android/app/build.gradle
@@ -3,7 +3,7 @@ apply plugin: 'com.android.application'
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "com.hbw.shortvideo"
+        applicationId "io.agora.MultiChannelSwitch"
         minSdkVersion 16
         targetSdkVersion 28
         versionCode 2
@@ -23,22 +23,20 @@ android {
 }
 
 dependencies {
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
+
     implementation 'androidx.appcompat:appcompat:1.0.2'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.0'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
 
-    def agora_sdk_version = "4.1.0-1"
-    implementation "io.agora.rtc:full-sdk:${agora_sdk_version}"
-
-    // 阿里播放器
-    implementation (name:'AliyunPlayer-4.7.2-full',ext:'aar')
-    implementation 'com.alivc.conan:AlivcConan:0.9.5'
+    def agora_sdk_version = "4.1.0"
+    // implementation "io.agora.rtc:full-sdk:${agora_sdk_version}"
+    implementation "io.agora.rtc:full-rtc-basic:${agora_sdk_version}"
     // 垂直方向的viewpager
     implementation 'com.github.castorflex.verticalviewpager:library:19.0.1'
     // 图片加载框架glide
     implementation 'com.github.bumptech.glide:glide:4.8.0'
     annotationProcessor 'com.github.bumptech.glide:compiler:4.8.0'
+
 }
