+++ Android/app/src/main/java/io/agora/MultiChannelSwitch/MainActivity.java
@@ -3,56 +3,37 @@ package io.agora.MultiChannelSwitch;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.viewpager.widget.PagerAdapter;
 import androidx.viewpager.widget.ViewPager;
 
 import android.os.Bundle;
-import android.text.TextUtils;
 import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.SurfaceHolder;
-import android.view.SurfaceView;
 import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.LinearLayout;
-import android.widget.TextView;
-
-import com.aliyun.player.AliPlayer;
-import com.aliyun.player.AliPlayerFactory;
-import com.aliyun.player.IPlayer;
-import com.aliyun.player.nativeclass.CacheConfig;
-import com.aliyun.player.source.UrlSource;
-import com.bumptech.glide.Glide;
-import com.hbw.shortvideo.R;
 
 import io.agora.MultiChannelSwitch.common.AgoraManager;
 import io.agora.MultiChannelSwitch.common.HostInfo;
 import io.agora.MultiChannelSwitch.common.HostVideoInfo;
-import io.agora.MultiChannelSwitch.common.VideoReportLayout;
 
 import java.util.ArrayList;
-import java.util.LinkedList;
 import java.util.List;
 
 import fr.castorflex.android.verticalviewpager.VerticalViewPager;
-import io.agora.rtc2.RtcConnection;
+
+import io.agora.MultiChannelSwitch.libs.VerticalPageAdapter;
 import io.agora.rtc2.RtcEngineConfig;
 
 public class MainActivity extends AppCompatActivity {
 
     private VerticalViewPager vp;
 
-    private AgoraPageAdapter agoraPagerAdapter;
-    private int currentFlagPostion;//传递过来播放第几个
+    private VerticalPageAdapter verticalPagerAdapter;
+    private int currentFlagPosition;//传递过来播放第几个
 
     private int mCurrentPosition;//当前正在播放第几个主播
-    private int mPrevousPosition = -1;//上一个播放的主播
-
-    private List<HostInfo> hosts = new ArrayList<>();
+    private int mPreviousPosition = -1;//上一个播放的主播
 
     private void initSampleHosts()
     {
+        List<HostInfo> hosts = new ArrayList<>();
         int hostCount = 3; // set number to 12 for loop
 
         String[] channels = new String[] {"AoVf4Hfdy9EbyDLh", "RxhMtitAzUwk1nBR", "IBbgSrJic1gNbGaG"};
@@ -62,6 +43,8 @@ public class MainActivity extends AppCompatActivity {
         {
             hosts.add(new HostInfo(channels[i%3], 900001, images[i%3]));
         }
+
+        AgoraManager.getInstance().setHosts(hosts);
     }
 
     @Override
@@ -96,44 +79,48 @@ public class MainActivity extends AppCompatActivity {
     }
 
     private void initViewPager() {
-        currentFlagPostion = getIntent().getIntExtra("currentPostion", 0);
+        currentFlagPosition = getIntent().getIntExtra("currentPosition", 0);
 
-        agoraPagerAdapter = new AgoraPageAdapter();
+        verticalPagerAdapter = new VerticalPageAdapter();
 
-        vp.setAdapter(agoraPagerAdapter);
+        vp.setAdapter(verticalPagerAdapter);
         vp.setOffscreenPageLimit(1);
         vp.setOnPageChangeListener(new ViewPager.OnPageChangeListener() {
             @Override
             public void onPageScrolled(int i, float v, int i1) {
 //                Log.i("PageScrolling", "onPageScrolled i:"+i+" v:"+v+" i1:"+i1);
+                // TODO: Join Channel based on Scrolling position
+
+
             }
 
             @Override
             public void onPageSelected(int position) {
 
-                mPrevousPosition = mCurrentPosition;
+                mPreviousPosition = mCurrentPosition;
 
                 /**
-                 * mute prevous channel
+                 * mute previous channel
                  * unsubscribe any hosts
                  * leave channel
                  */
-                HostVideoInfo hostVideoInfo = agoraPagerAdapter.findHostVideoInfo(mPrevousPosition);
-                if(hostVideoInfo != null && hostVideoInfo.isJoined()) {
-                    Log.i("DEMO", "mute Previous Host Audio:"+mPrevousPosition);
+                HostVideoInfo hostVideoInfo = verticalPagerAdapter.findHostVideoInfo(mPreviousPosition);
+                if(hostVideoInfo != null) {
+                    Log.i("DEMO", "mute Previous Host Audio:"+mPreviousPosition);
                     AgoraManager.getInstance().unMuteAudioEx(hostVideoInfo.getConnection(), true);
-                    Log.i("DEMO", "unSubscribe Channel:"+mPrevousPosition);
+                    Log.i("DEMO", "unSubscribe Channel:"+mPreviousPosition);
                     AgoraManager.getInstance().updateChannelEx(hostVideoInfo.getConnection(), false);
-                    Log.i("DEMO", "Leave channel:"+mPrevousPosition);
+
 
                     // Here, call agora leave channel function is not a good idea,
                     // because if user swipe back to previous one
-                    AgoraManager.getInstance().leaveChannelEx(hostVideoInfo.getConnection());
-                    hostVideoInfo.setJoined(false);
+                    // AgoraManager.getInstance().leaveChannelEx(hostVideoInfo.getConnection());
+                    // hostVideoInfo.setJoined(false);
+                    // Log.i("DEMO", "Leave channel:"+mPreviousPosition);
                 }
 
                 mCurrentPosition = position;
-                Log.i("DEMO", "onPageSelected position:"+position);
+                //Log.i("DEMO", "onPageSelected position:"+position);
 
             }
 
@@ -145,10 +132,30 @@ public class MainActivity extends AppCompatActivity {
         vp.setPageTransformer(false, new ViewPager.PageTransformer() {
             @Override
             public void transformPage(@NonNull View page, float position) {
+//                if (position < -1) {
+//                    HostVideoInfo nextHostInfo = agoraPagerAdapter.findHostVideoInfo(mCurrentPosition + 1);
+//
+//                    if(nextHostInfo != null && !nextHostInfo.isJoined()) {
+//                        Log.i("DEMO", "Next Host state:"+nextHostInfo.isJoined());
+//                        AgoraManager.getInstance().joinChannelEx(nextHostInfo.getConnection(), "");
+//                        nextHostInfo.setJoined(true);
+//                    }
+//                }
+//
+//                if (position > 1 ) {
+//                    HostVideoInfo prevHostInfo = agoraPagerAdapter.findHostVideoInfo(mCurrentPosition - 1);
+//
+//                    if(prevHostInfo != null && !prevHostInfo.isJoined()) {
+//                        Log.i("DEMO", "Prev Host state:"+prevHostInfo.isJoined());
+//                        AgoraManager.getInstance().joinChannelEx(prevHostInfo.getConnection(), "");
+//                        prevHostInfo.setJoined(true);
+//                    }
+//                }
                 if (position != 0) {
+//                    Log.i("POSITION", "transformPage position:"+position);
                     return;
                 }
-                Log.i("DEMO", "transformPage position:"+mCurrentPosition);
+//                Log.i("DEMO", "transformPage position:"+mCurrentPosition);
 
                 /**
                  * Here is the place when NEW page is fully displayed.
@@ -156,11 +163,14 @@ public class MainActivity extends AppCompatActivity {
                  * So, subscribe channel host and unMute audio
                  * Setup display UI
                  */
-
-                HostVideoInfo hostVideoInfo = agoraPagerAdapter.findHostVideoInfo(mCurrentPosition);
+                HostVideoInfo hostVideoInfo = verticalPagerAdapter.findHostVideoInfo(mCurrentPosition);
                 if(hostVideoInfo != null) {
+
+                    Log.i("DEMO", "subscribe Channel:"+mCurrentPosition + " " +hostVideoInfo.getHost().getChannelId());
                     AgoraManager.getInstance().updateChannelEx(hostVideoInfo.getConnection(), true);
+                    Log.i("DEMO", "unmute Host Audio:"+mCurrentPosition);
                     AgoraManager.getInstance().unMuteAudioEx(hostVideoInfo.getConnection(), false);
+
                     AgoraManager.getInstance().setUidViewEx(hostVideoInfo.getConnection()
                             , hostVideoInfo.getHost().getUserId(), hostVideoInfo.getHostVideo());
 
@@ -169,121 +179,8 @@ public class MainActivity extends AppCompatActivity {
             }
 
         });
-        vp.setCurrentItem(currentFlagPostion);
+        vp.setCurrentItem(currentFlagPosition);
     }
-
-    class AgoraPageAdapter extends PagerAdapter {
-        private LinkedList<View> mViewCache = new LinkedList<>();
-        ArrayList<HostVideoInfo> hostVideoList = new ArrayList<>();
-        @Override
-        public boolean isViewFromObject(@NonNull View view, @NonNull Object o) {
-            return view == o;
-        }
-
-        @NonNull
-        @Override
-        public Object instantiateItem(@NonNull ViewGroup container, final int position) {
-            View view;
-//            if (mViewCache.size() == 0) {
-                view = LayoutInflater.from(container.getContext()).inflate(R.layout.fragment_single_host, null, false);
-//            } else {
-//                view = mViewCache.removeFirst();
-//            }
-            view.setId(position);
-
-            HostInfo host = hosts.get(position % 3);
-
-            Log.i("instantiateItem", "instantiateItem position:"+position);
-            Log.i("DEMO", "Join Channel:"+position);
-
-            TextView videoTitle = view.findViewById(R.id.hostTitle);
-            final ImageView coverPicture = view.findViewById(R.id.hostImage);
-            videoTitle.setText("<  " + host.getChannelName());
-
-            /**
-             * This is for Internet Images
-             */
-//            if (!TextUtils.isEmpty(host.getImageUrl())) {
-//                coverPicture.setVisibility(View.VISIBLE);
-//                Glide.with(MainActivity.this).load(host.getImageUrl()).into(coverPicture);
-//            }
-
-            /**
-             * Local images
-             */
-            coverPicture.setVisibility(View.VISIBLE);
-            coverPicture.setImageResource(host.getImageId());
-
-
-            VideoReportLayout video = view.findViewById(R.id.hostVideo);
-
-            HostVideoInfo hostInfo = new HostVideoInfo();
-            hostInfo.setHost(host);
-            hostInfo.setHostVideo(video);
-            hostInfo.setPosition(position);
-
-            RtcConnection connection = new RtcConnection();
-            connection.channelId = host.getChannelName();
-            connection.localUid = 10002;
-
-            hostInfo.setConnection(connection);
-
-            AgoraManager.getInstance().joinChannelEx(connection, "");
-
-            hostInfo.setJoined(true);
-
-            hostVideoList.add(hostInfo);
-
-            container.addView(view);
-            return view;
-        }
-
-        @Override
-        public int getCount() {
-            return hosts.size();
-        }
-
-        @Override
-        public void destroyItem(@NonNull ViewGroup container, int position, @NonNull Object object) {
-            /**
-             * Perhaps, it's a good idea that
-             * Here is the good place to call leave channel method
-             */
-            //leaveHostChannel(position);
-
-            View contentView = (View) object;
-            container.removeView(contentView);
-            //mViewCache.add(contentView);
-            Log.i("DEMO", "Destroy item :"+position);
-
-        }
-
-        public HostVideoInfo findHostVideoInfo(int position) {
-            if( position < 0 ) return null;
-            for (int i = 0; i < hostVideoList.size(); i++) {
-                HostVideoInfo hostInfo = hostVideoList.get(i);
-                if (hostInfo.getPosition() == position) {
-                    return hostInfo;
-                }
-            }
-            return null;
-        }
-
-        protected void leaveHostChannel(int position) {
-
-            HostVideoInfo hostInfo = findHostVideoInfo(position);
-            if (hostInfo == null)
-                return;
-
-            AgoraManager.getInstance().leaveChannelEx(hostInfo.getConnection());
-            Log.i("DEMO", "Leave channel :" + position);
-            hostInfo.setJoined(false);
-
-        }
-
-    }
-
-
 
     @Override
     protected void onDestroy() {
