+++ Android/app/src/main/res/layout/fragment_single_host.xml
@@ -5,6 +5,15 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
+
+    <ImageView
+        android:id="@+id/hostImage"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_gravity="top"
+        android:scaleType="fitXY"
+        android:visibility="gone" />
+
     <TextView
         android:id="@+id/hostTitle"
         android:layout_width="match_parent"
@@ -13,25 +22,15 @@
         android:gravity="center_vertical"
         android:paddingLeft="20dp"
         android:textColor="@color/colorWhite"
-        android:text="tetet"
+        android:text=""
         android:textSize="16sp"/>
 
-    <ImageView
-        android:id="@+id/hostImage"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_gravity="center"
-        android:scaleType="fitCenter"
-        android:visibility="gone" />
-
-    <io.agora.MultiChannelSwitch.common.VideoReportLayout
+    <io.agora.MultiChannelSwitch.common.VideoLayout
         android:id="@+id/hostVideo"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:focusable="true"
         android:visibility="visible"
         tools:ignore="ExtraText" />
-
-
 
 </FrameLayout>
